refactor(api): move user-repository (back) to enrolment
Previously moved to shared in PR #10404, no longer justified
