+++ api/src/certification/enrolment/infrastructure/repositories/index.js
@@ -3,7 +3,6 @@ import { injectDependencies } from '../../../../shared/infrastructure/utils/depe
 import * as sessionManagementRepository from '../../../session-management/infrastructure/repositories/session-management-repository.js';
 import * as certificationCenterRepository from '../../../shared/infrastructure/repositories/certification-center-repository.js';
 import * as targetProfileHistoryRepository from '../../../shared/infrastructure/repositories/target-profile-history-repository.js';
-import * as userRepository from '../../../shared/infrastructure/repositories/user-repository.js';
 import boundedContext from '../../dependencies.json' with { type: 'json' };
 import * as candidateRepository from './candidate-repository.js';
 import * as certificationCpfCityRepository from './certification-cpf-city-repository.js';
@@ -13,6 +12,7 @@ import * as complementaryCertificationCourseRepository from './complementary-cer
 import * as divisionRepository from './division-repository.js';
 import * as sessionForAttendanceSheetRepository from './session-for-attendance-sheet-repository.js';
 import * as sessionRepository from './session-repository.js';
+import * as userRepository from './user-repository.js';
 
 /**
  * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
