+++ api/tests/certification/enrolment/integration/infrastructure/repositories/user-repository_test.js
@@ -1,4 +1,4 @@
-import * as userRepository from '../../../../../../src/certification/shared/infrastructure/repositories/user-repository.js';
+import * as userRepository from '../../../../../../src/certification/enrolment/infrastructure/repositories/user-repository.js';
 import { expect } from '../../../../../test-helper.js';
 import { databaseBuilder } from '../../../../../tooling/databases.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
